authentication

+++ src/main/java/com/cardosojl/repositories/MerchantRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cardosojl.models.Merchant;
 
 @Repository
-public interface MerchantRepository extends JpaRepository<Merchant, Integer>{
+public interface MerchantRepository extends JpaRepository<Merchant, Long>{
 
 }
